// frontend/lib/api/model/searchExercisesApiV1ExercisesGetParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Workout API
 * A workout tracking API built with FastAPI
 * OpenAPI spec version: 1.0.0
 */

export type SearchExercisesApiV1ExercisesGetParams = {
  /**
   * Filter by exercise name (partial match)
   */
  name?: string | null;
  /**
   * Filter by body part
   */
  body_part?: string | null;
  /**
   * Filter by exercise modality
   */
  modality?: string | null;
  /**
   * Filter by user-created vs system exercises
   */
  is_user_created?: boolean | null;
  /**
   * Filter exercises created by specific user
   */
  created_by_user_id?: number | null;
  /**
   * Page number (1-based)
   */
  page?: number;
  /**
   * Page size (max 100)
   */
  size?: number;
};