// frontend/lib/api/model/pageExerciseResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Workout API
 * A workout tracking API built with FastAPI
 * OpenAPI spec version: 1.0.0
 */
import type { ExerciseResponse } from "./exerciseResponse";

export interface PageExerciseResponse {
  items: ExerciseResponse[];
  /** Current page number */
  page: number;
  /** Total number of pages */
  pages: number;
  /** Page size */
  size: number;
  /** Total number of items */
  total: number;
}